feat: extract common fn for notify
